requirements.txt
fastapi
uvicorn
python-multipart
requests 
python-jose
passlib